CI as a GitHub Action
There is several fixes and workarounds which need to be done before running the test file:

The some files in GRASS GIS doc are not installed properly, so we get them online to make g.extension work. --tmp-location is in 7.6 but not 7.4, so we create a location manually. Similarly, --tmp-mapset will be only in 8.0, so we create and delete one manually (in case we want to cache the location later). Shebang for 7.8 and above now should be python3, but 7.4 is Python 2, so we change it to just python.

The test call itself is directly calling the file which is allowed, but might not be ideal for more advanced tests.

We are assuming that the name of the repository is the name of the module.